editar tiles layer consertado, tilesetandobjects panel funcionando

+++ src/TileMap.cpp
@@ -120,8 +120,8 @@ void TileMap::renderLayer(int layer, int cameraX, int cameraY)
 					case 0:
 						tileSet_->render( 
 							(unsigned)at(j, i, layer), 
-							(float)(j * tileSet_->getTileWidth()  + cameraX - Camera::getPosition().x() * 0.5), 
-							(float)(i * tileSet_->getTileHeight() + cameraY - Camera::getPosition().y() * 0.5) 
+							(float)(j * tileSet_->getTileWidth()  + cameraX - Camera::getPosition().x() * (double)0.5), 
+							(float)(i * tileSet_->getTileHeight() + cameraY - Camera::getPosition().y() * (double)0.5) 
 						);
 						break;
 					case 1:
@@ -134,15 +134,15 @@ void TileMap::renderLayer(int layer, int cameraX, int cameraY)
 					case 2:
 						tileSet_->render( 
 							(unsigned)at(j, i, layer), 
-							(float)(j * tileSet_->getTileWidth()  + cameraX + Camera::getPosition().x() * 0.5), 
-							(float)(i * tileSet_->getTileHeight() + cameraY + Camera::getPosition().y() * 0.5) 
+							(float)(j * tileSet_->getTileWidth()  + cameraX + Camera::getPosition().x() * (double)0.5), 
+							(float)(i * tileSet_->getTileHeight() + cameraY + Camera::getPosition().y() * (double)0.5) 
 						);
 						break;
 					case 3:
 						tileSet_->render( 
 							(unsigned)at(j, i, layer), 
-							(float)(j * tileSet_->getTileWidth()  + cameraX + Camera::getPosition().x() * 0.75), 
-							(float)(i * tileSet_->getTileHeight() + cameraY + Camera::getPosition().y() * 0.75) 
+							(float)(j * tileSet_->getTileWidth()  + cameraX + Camera::getPosition().x() * (double)0.75), 
+							(float)(i * tileSet_->getTileHeight() + cameraY + Camera::getPosition().y() * (double)0.75) 
 						);
 						break;
 					default:
